+++ portal-ui/src/screens/Console/Configurations/SiteReplication/AddReplicationSites.tsx
@@ -35,15 +35,7 @@ import { useAppDispatch } from "../../../../store";
 import { useSelector } from "react-redux";
 import { selSession } from "../../consoleSlice";
 import SRSiteInputRow from "./SRSiteInputRow";
-
-export type SiteInputRow = {
-  name: string;
-  endpoint: string;
-  accessKey: string;
-  secretKey: string;
-  isCurrent?: boolean;
-  isSaved?: boolean;
-};
+import { SiteInputRow } from "./Types";
 
 const isValidEndPoint = (ep: string) => {
   let isValidEndPointUrl = false;
@@ -489,7 +481,7 @@ const AddReplicationSites = () => {
                       display: "flex",
                       marginTop: "15px",
                       marginBottom: "15px",
-                      width:"100%",
+                      width: "100%",
 
                       "&.step-text": {
                         fontWeight: 400,
@@ -509,8 +501,8 @@ const AddReplicationSites = () => {
                   <Box>
                     The following changes are replicated to all other sites
                   </Box>
-                  <ul >
-                    <li >Creation and deletion of buckets and objects</li>
+                  <ul>
+                    <li>Creation and deletion of buckets and objects</li>
                     <li>
                       Creation and deletion of all IAM users, groups, policies
                       and their mappings to users or groups
@@ -521,42 +513,52 @@ const AddReplicationSites = () => {
                       owned by the root user)
                     </li>
                     <li>
-                        <Box  style={{
-                            display:"flex",
-                            flexFlow:"column",
+                      <Box
+                        style={{
+                          display: "flex",
+                          flexFlow: "column",
 
-                            justifyContent:"flex-start",
-                        }}>
-                            <div style={{
-                                paddingTop:"1px",
-                            }}>Changes to Bucket features such as</div>
-                      <ul>
-
-                        <li>Bucket Policies</li>
-                        <li>Bucket Tags</li>
-                        <li>Bucket Object-Lock configurations</li>
-                        <li>Bucket Encryption configuration</li>
-                      </ul>
-                        </Box>
+                          justifyContent: "flex-start",
+                        }}
+                      >
+                        <div
+                          style={{
+                            paddingTop: "1px",
+                          }}
+                        >
+                          Changes to Bucket features such as
+                        </div>
+                        <ul>
+                          <li>Bucket Policies</li>
+                          <li>Bucket Tags</li>
+                          <li>Bucket Object-Lock configurations</li>
+                          <li>Bucket Encryption configuration</li>
+                        </ul>
+                      </Box>
                     </li>
 
                     <li>
-                        <Box  style={{
-                            display:"flex",
-                            flexFlow:"column",
+                      <Box
+                        style={{
+                          display: "flex",
+                          flexFlow: "column",
 
-                            justifyContent:"flex-start",
-                        }}>
-                            <div style={{
-                                paddingTop:"1px",
-                            }}>The following Bucket features will NOT be replicated</div>
+                          justifyContent: "flex-start",
+                        }}
+                      >
+                        <div
+                          style={{
+                            paddingTop: "1px",
+                          }}
+                        >
+                          The following Bucket features will NOT be replicated
+                        </div>
 
-
-                      <ul>
-                        <li>Bucket notification configuration</li>
-                        <li>Bucket lifecycle (ILM) configuration</li>
-                      </ul>
-                        </Box>
+                        <ul>
+                          <li>Bucket notification configuration</li>
+                          <li>Bucket lifecycle (ILM) configuration</li>
+                        </ul>
+                      </Box>
                     </li>
                   </ul>
                 </Box>
