+++ portal-ui/src/screens/Console/Configurations/SiteReplication/Types.tsx
@@ -1,0 +1,8 @@
+export type SiteInputRow = {
+  name: string;
+  endpoint: string;
+  accessKey: string;
+  secretKey: string;
+  isCurrent?: boolean;
+  isSaved?: boolean;
+};
